Added sign up page
In the process of adding a user login system.Already added a sign up
page. was planning to redirect to an account created page after sign
up but will probably scrap that idea and resort to a flash message
instead. The user is not being saved to database as of now, only
validation takes place when one hits sign up.